vertex_data: solidified the response to numerical errors from the lp problem
If the weights are very small, essentially the present form of rewriting the detailed balance equations becomes unstable. The real solution would be to solve the equations directly. For now we truncate to the closest value that makes sense and print a warning that this might be inaccurate. For practical situations, this hopefully is not ever important.