dump_format: change opercode storage type
Since the internal opercode int type was enlarged, the dumps have been broken. Now we change the stored int format to be the same to fix that. As a result, we have dump incompatibility across all models. Maybe this should be softened later by allowing to set the int type at compile time.